Moved other classes to query package
