+++ genomics_data_index/__init__.py
@@ -1,3 +1,3 @@
-from genomics_data_index.api.GenomicsDataIndex import GenomicsDataIndex
+from genomics_data_index.api.query.GenomicsDataIndex import GenomicsDataIndex
 
 __version__: str = '0.1.0.dev1'
