+++ genomics_data_index/api/query/GenomicsDataIndex.py
@@ -7,7 +7,7 @@ from typing import List
 
 import pandas as pd
 
-from genomics_data_index.api.SamplesQuery import SamplesQuery
+from genomics_data_index.api.query.SamplesQuery import SamplesQuery
 from genomics_data_index.api.query.impl.DataFrameSamplesQuery import DataFrameSamplesQuery
 from genomics_data_index.api.query.impl.SamplesQueryIndex import SamplesQueryIndex
 from genomics_data_index.api.query.impl.TreeSamplesQuery import TreeSamplesQuery
