+++ genomics_data_index/test/integration/api/test_GenomicsDataIndex_samples_query.py
@@ -3,8 +3,8 @@ import math
 import pandas as pd
 import pytest
 
-from genomics_data_index.api.GenomicsDataIndex import GenomicsDataIndex
-from genomics_data_index.api.SamplesQuery import SamplesQuery
+from genomics_data_index.api.query.GenomicsDataIndex import GenomicsDataIndex
+from genomics_data_index.api.query.SamplesQuery import SamplesQuery
 from genomics_data_index.api.query.impl.TreeSamplesQuery import TreeSamplesQuery
 from genomics_data_index.configuration.connector.DataIndexConnection import DataIndexConnection
 from genomics_data_index.test.integration import snippy_all_dataframes, data_dir
